requirements.txt
alembic==1.10.4
fastapi~=0.104.1
psycopg2-binary==2.9.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.2
python-multipart
libgravatar==1.0.4
fastapi-mail
redis~=4.6.0
fastapi-limiter==0.1.5
cloudinary==1.32.0
sqlalchemy~=2.0.23
pydantic~=2.4.2
starlette~=0.27.0
uvicorn~=0.24.0.post1
streamlit~=1.28.1
PyPDF2~=3.0.1
langchain~=0.0.335
faiss-cpu==1.7.4
openai
tiktoken
requests~=2.31.0
pydantic[dotenv]
unstructured
ebooklib
python-pptx
python-docx

PyJWT~=2.8.0
python-dotenv~=1.0.0
docx~=0.2.4

textract~=1.6.5
ipywidgets~=8.1.1
transformers~=4.35.2
pandas~=2.1.3
ipython~=8.17.2
matplotlib~=3.8.1